tracking/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
python-osc>=1.8.0
termcolor>=2.3.0
pydantic>=2.0.0
pyyaml>=6.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
psutil>=5.9.0
aiohttp>=3.8.0
typer>=0.9.0
rich>=13.0.0
pynvml>=11.5.0
